terminando de documentar

+++ src/main/java/pe/edu/utp/sistemadereservacionhotel/service/patron/creacional/ComprobanteFactory.java
@@ -3,6 +3,8 @@ package pe.edu.utp.sistemadereservacionhotel.service.patron.creacional;
 import pe.edu.utp.sistemadereservacionhotel.model.finanzas.ComprobantePago;
 import pe.edu.utp.sistemadereservacionhotel.model.reserva.Reserva;
 
+import java.math.BigDecimal;
+
 public interface ComprobanteFactory {
-    ComprobantePago crear(Reserva reserva, Double montoTotal);
+    ComprobantePago crear(Reserva reserva, BigDecimal montoTotal);
 }
